+++ src/onegov.gazette/onegov/gazette/fields.py
@@ -1,6 +1,16 @@
 from onegov.form.fields import MultiCheckboxField as MultiCheckboxFieldBase
+from onegov.form.widgets import MultiCheckboxWidget as MultiCheckboxWidgetBase
 from onegov.gazette import _
 from wtforms.fields.html5 import DateTimeLocalField as DateTimeLocalFieldBase
+
+
+class MultiCheckboxWidget(MultiCheckboxWidgetBase):
+
+    def __call__(self, field, **kwargs):
+        kwargs['data-expand-title'] = field.gettext(_("Show all"))
+        kwargs['data-fold-title'] = field.gettext(_("Show less"))
+
+        return super(MultiCheckboxWidgetBase, self).__call__(field, **kwargs)
 
 
 class MultiCheckboxField(MultiCheckboxFieldBase):
@@ -10,22 +20,14 @@ class MultiCheckboxField(MultiCheckboxFieldBase):
     Also, disables all the options if the whole field is disabled.
     """
 
+    widget = MultiCheckboxWidget()
+
     def __init__(self, *args, **kwargs):
         render_kw = kwargs.pop('render_kw', {})
-        render_kw['data-limit'] = str(kwargs.pop('limit', 10))
-        render_kw['data-expand-title'] = _("Show all")
-        render_kw['data-fold-title'] = _("Show less")
+        render_kw['data-limit'] = str(kwargs.pop('limit', 5))
         kwargs['render_kw'] = render_kw
 
         super().__init__(*args, **kwargs)
-
-    def translate(self, request):
-        self.render_kw['data-expand-title'] = request.translate(
-            self.render_kw['data-expand-title']
-        )
-        self.render_kw['data-fold-title'] = request.translate(
-            self.render_kw['data-fold-title']
-        )
 
     def __iter__(self):
         for opt in super(MultiCheckboxField, self).__iter__():