+++ src/onegov.gazette/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Uses the latest onegov.form.
   [msom]
 
+- Uses a widget to translate the multi checkbox fields.
+  [msom]
+
 1.17.2 (2018-10-02)
 ~~~~~~~~~~~~~~~~~~~~